[FIX]ForEachPromise delete
나만의 forEach promise를 만들려했지만 그 방법보다 promise.all 이용한뒤에 reduce를 이용하는 것이 더 괜찮은 방법인 것 같아서 방법을 바꿈
